requirements.txt
opencv-python
pillow
google-generativeai
#google-genai
python-dotenv
pywhatkit
pywin32
flask